// frontend/src/types/api.generated.ts
export interface components {
    schemas: {
// ...
        ChangeEmailRequest: {
            CurrentPassword: string;
            /** Format: email */
            NewEmail: string;
        };
        ChangePasswordRequest: {
            CurrentPassword: string;
            NewPassword: string;
            ConfirmNewPassword: string;
        };
// ...
        LoginRequest: {
            Username: string;
            Password: string;
        };
        RefreshTokenRequest: {
            Username?: string | null;
            RefreshToken?: string | null;
        };
// ...
        UserProfileRequest: {
            FirstName?: string | null;
            LastName?: string | null;
            PhoneNumber?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}